cedar-go: run integration tests as part of the regular test suite
This change adds a local copy of the corpus-tests.tar.gz from the upstread cedar-policy/cedar-integration-tests repo and a test harness which loads the tests in-memory and runs them.

This will allow us to run these tests as part of our continuous integration pipeline and should reduce noise due to network blips. It will also mean that upstream changes to the formatting of the tests won't instantly break the test runner, leaving developers unsure if their code is correct or not.

The tests run in just over 1 second on my dev machine, so I see no harm in running them on every PR.

Additionally, the nightly test corpus run has been transformed into a check for a difference in the upstream version and the local version of the tests. If a change happens upstream, an issue will be opened which we can use to track updating the local copy of the tests.

Signed-off-by: Patrick Jakubowski <patrick.jakubowski@strongdm.com>