This commit adds an extension type, datetime, to Cedar Go.
This PR does not add the duration type, nor the methods `toTime()`,
`offset()`, or `durationSince()` that would require it.

This PR adds operator overloading, by introducing a set of "Virtual"
comparison Evaler. Types that can overload operators must implement
the `Lesser` interface. Long and Datetime are implemented in this
PR. (Note: It is expected that there will be a Decimal operator
overload RFC at some later date, but that is not hooked into this)

The now obsolete Evalers for Long comparisons have not been
removed. A possible improvement to this PR would be to utilize
them when it is known that Longs are on the right and left side.

The datetime itself is backed by an int64, and can be lifted from
a Go `time.Time`, integer or float "unsafely."

Signed-off-by: Andrew Gwozdziewycz <andrew.gwozdziewycz@strongdm.com>