cedar-go: use EntityGetter interface in PolicySet.IsAuthorized()
This replaces the concrete EntityMap passed to IsAuthorized() with an
EntityGetter interface, allowing alternative implementations of this interface
to be supplied for authorization.

While this is hypothetically a breaking change, we think that is extremely
unlikely to be the case in practice. Any code passing an EntityMap will
continue to work - the only potential breakage would be if a caller were
directly passing a map[EntityUID]Entity here.

Signed-off-by: Kevin Jamieson <kevin.jamieson@strongdm.com>