cedar-go: clarify current policy around backwards compatibility
At present time, the maintainers are reserving the right to:
* Add variadics to existing functions/methods.
* Replace concrete types with interfaces in the parameters to functions/methods
  (as occurred in this patch set).

While strictly breaking changes, we are asserting that the benefits of being
able to make changes of these forms without a major revision outweigh the
unlikely possibility of breaking code written against cedar-go.

We may of course choose to revisit and revise these policies in the future,
especially as cedar-go gains wider usage.

Signed-off-by: Kevin Jamieson <kevin.jamieson@strongdm.com>