Addresses cedar-policy/rfcs#94 concerns about time wrapping
The Go time package will silently ignore dates that don't exist on the
calendar, instead carrying forward the erroneous time component. As an
example, `2024-02-30` will become `2024-03-01`, since there was no
February 30th, in 2024 (or ever).

This change returns an error when there's a discrepancy between the
numbers we parse from the string, and the component after creating a
Go time.Time value from the parsed numbers.

Signed-off-by: Andrew Gwozdziewycz <andrew.gwozdziewycz@strongdm.com>