cedar-go: Add a new top-level Authorize() function
This function is identical to PolicySet.IsAuthorized(), but allows a generic iterator of (PolicyID, *Policy) tuples to be passed. This allows for flexibility at the callsite if the caller wants to use some other data structure to hold on to a map of PolicyIDs to *Policy.

Note: an All() method is also added to PolicySet which returns an iterator over the internal policy map.

Signed-off-by: Patrick Jakubowski <patrick.jakubowski@strongdm.com>