cedar-go: upgrade golangci-lint checks to v2 and fix all issues found
I preserved the list of linters that we run, but enabled almost all of the revive rules. We can disable them as we see fit in the future.

The only regrettable exclusion is on the `exported` rule in the `types/` directory. We should make an effort to document all of the public types.

Signed-off-by: Patrick Jakubowski <patrick.jakubowski@strongdm.com>