cedar-go: remove vendor directory
Having a vendor directory for a library module like the SDK is of dubious value generally. Ultimately, the library will be consumed by an executable binary whose final set of dependencies will be determined by looking at all of its transitive library dependencies and our set of vendored dependencies will be ignored in that process.

I did a survey of a few dozen of the most popular Go libraries and couldn't find another instance of an library module vendoring its dependencies.

Signed-off-by: Patrick Jakubowski <patrick.jakubowski@strongdm.com>