Add checksumtype to the golangci-lint file.
Some of the types within cedar-go match the sum type idiom (an interface with an unexported method) described in the linter below.

Let's add an explicit check for exhaustiveness for this specific linter.

https://github.com/alecthomas/go-check-sumtype

*Issue #, if available:*

IDK what issue tracker you use.

*Description of changes:*

Add a sum type annotation, add two branches to convertNamespace, both of which are just comments.
Let me know if the behavior in those branches should be a little smarter.

Signed-off-by: Greg NISBET <gregory.nisbet@gmail.com>